requirements.txt
django==1.11.1
django-oauth-toolkit
django-cors-middleware
djangorestframework